% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CountryPopulation.R
\name{CountryPopulation}
\alias{CountryPopulation}
\title{Generate a graph of a country's estimated population in thousands from 1950
to 2020.}
\usage{
CountryPopulation(x)
}
\arguments{
\item{x}{A string}
}
\value{
a graph of the population of x
}
\description{
Generate a graph of a country's estimated population in thousands from 1950
to 2020. The function takes a string as input. if the string matches a country name
in the data frame `WorldPopulation`, the function will produce a line graph
of estimated population data from the United Nations.
}
\examples{
CountryPopulation('China')
CountryPopulation('United States of America')
}
