% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/World_Cup.R
\docType{data}
\name{World_Cup}
\alias{World_Cup}
\title{Attendance information for every World Cup.}
\format{
a data frame with 22 observations of 4 variables
\describe{
  \item{Matches}{Number of soccer matches}
  \item{Totalattendance}{total number of attendees across all matches}
  \item{Averageattendance}{average number of attendees per match}
  \item{WorldCup}{Country name and world cup year without spaces formatted
  as a factor}
}
}
\source{
\url{https://en.wikipedia.org/wiki/FIFA_World_Cup#Attendance}
}
\usage{
World_Cup
}
\description{
Attendance information for every World Cup.
}
\keyword{datasets}
